backend/requirements.txt
websockets==12.0
discord.py==2.3.2
anthropic==0.96.0
python-dotenv==1.0.0
